+++ source/dijkstra.cpp
@@ -2,7 +2,7 @@
 
 Dijkstra::Dijkstra() : Astar(0, CN_SP_BT_GMAX) {}
 
-double Dijkstra::computeHFromCellToCell(int i1, int j1, int i2, int j2, const EnvironmentOptions &options) {
+double Dijkstra::computeHFromCellToCell(double i1, double j1, double i2, double j2, const EnvironmentOptions &options) {
     return 0;
 }
