+++ source/dijkstra.h
@@ -7,6 +7,6 @@ class Dijkstra : public Astar
     public:
         Dijkstra();
     protected:
-        double computeHFromCellToCell(int i1, int j1, int i2, int j2, const EnvironmentOptions &options) override;
+        double computeHFromCellToCell(double i1, double j1, double i2, double j2, const EnvironmentOptions &options) override;
 };
 #endif
